v1.76 Fixes and tweaks as listed: 1: FOR RETRY - When someone joins a tribe by connecting to a multiplayer server, automatically being asked to join a tribe, and clicking the name of a tribe to join that tribe, the GUI doesn't close until you select the tribe 2 times.
2: DONE - When using "/tribe help", I'd like the wiki link to lead to:
https://github.com/Heltrato/tribes-mod-forge-1.18.2#readme
and I'd like the discord link to be removed.

3: DONE https://i.imgur.com/QClgHZb.jpeg
When a tribe leader joins a server and hasn't selected potion effects, hasn't promoted any members, and hasn't picked a deity to follow, the messages in that picture appear. I'd like to replace the 1st message in the top part of this image with:

Use "/tribe effects" to select your tribe's strengths & weaknesses.
I'd like the apostrophes and ` symbols in the middle message of that picture to be replaced with " quotation marks. If you see any other small typos like those where ` symbols or apostrophes are used in place of quotation marks, please fix those as well.

And I'd like to fix the last message by replacing 'tribe deity list' with "/tribe deity list" (including the quotation marks), adding the word "to" just before the word "see", replacing the word "options" with "a list", replacing "to follow." with "you can follow.", replacing the last 2 apostrophes with quotation marks, and adding a period after the last quotation marks.

4: DONE - When a tribe leader uses "/tribe hemisphere" without specifying which hemisphere they want to choose, the following message is displayed in white font:
pick which hemi to access (north / south)
I'd like that message to be replaced with this one and displayed in cyan font:
Pick which hemisphere (north or south) you want your tribe to be able to place and destroy blocks in. Your choice can never be changed, so choose carefully!

5: DONE - When the world is divided into eastern & western hemispheres via the main config file, "(north or south)" in the previous message should be replaced with "(east or west)"

6: DONE -  When a tribe leader tries to choose a hemisphere that their tribe can build in but their tribe's tier is too low to build outside of the strip of no man's land/wilderness in-between the 2 hemispheres, this message is displayed in cyan font:
Your tribe is not a high enough tier. Get more members
I'd like it to be replaced with this message, still in cyan font:
Your tribe isn't a high enough tier to access 1 of the 2 hemispheres. Use "/tribe invite send PlayerNameHere" to recruit more members, which will increase your tribe's tier and unlock benefits.

7: NEEDS A REWORK - I became the leader of a tribe with 2 members, and I lowered the threshold to become a tier 2 tribe to 2 members, but the game still says my tribe isn't a high enough tier to claim a hemisphere to build in. I then lowered the threshold for tribes to become tier 2 to just 1 member, and this glitch still occurs.

8: NEEDS A REWORK - I changed the "removeInactiveAfterDays" setting to kick inactive players after 1 day, but the change never came into effect; no players were kicked.

9: DONE - When I use "/tribe ban" as a tribe leader, the following message is displayed in red:
Specify a player to effect
I'd like that message to be replaced with this one and displayed in red:
Specify a player to ban.

10: NEEDS A LOG AND VIDEO ON THIS - When I use "/tribe ban playerNameHere" as a tribe leader on an online member of my tribe, I get this message in red:
An unexpected error occurred trying to execute that command
The player was successfully kicked, but that message shouldn't appear. I'd like that glitch to be fixed and a period added to the end of that message in cases when it does appear.

11: NEEDS A LOG AND VIDEO ON THIS - When I use "/tribe bans" as a tribe leader, rather than displaying a list of players who've been banned from the tribe, it gives the same message as point #9. It should list the players who've been banned from the tribe in alphabetical order, paginated.

12: DONE - https://i.imgur.com/EjLT30X.jpeg
When a player types "/tribe leave", I'd like this message to say:
Type "/tribe confirm" to leave your tribe.

13: NEEDS A REWORK - https://i.imgur.com/BYzGzh7.jpeg
This message appeared when I, as a tribe leader, left my tribe, and there was still 1 other player in it. I successfully left the tribe and don't know what error occurred.

14: DONE - When a new player joins your tribe, a message appears in chat that ends without any punctuation, specifically "playerNameHere has joined your tribe". I'd like it to end with an exclamation mark, and any other such messages that end without punctuation to end with a period. For example, when using "/tribe initials newInitialsHere", the 2 different messages you receive for valid and invalid initials both end without punctuation
I'm okay with ending sentences without punctuation when chatting with people, but when it comes to mod messages I think it'd look better if every sentence ended with punctuatio